[REFACTOR] extract in-memory metrics repository
